requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
colorthief==0.2.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
